refactor(search-view): extract useResultPipeline hook
